prctl: cast arg2 to int in PR_SET_PTRACER handler
Kernel code does this when compares it with -1 (apart from comparing it with
PR_SET_PTRACER_ANY, which is also -1) in security/yama/yama_lsm.c.

* prctl.c (SYS_FUNC(prctl)) <PR_SET_PTRACER>: Cast arg2 to int before
comparing it with -1.

+++ prctl.c
@@ -255,7 +255,7 @@ SYS_FUNC(prctl)
 
 	case PR_SET_PTRACER:
 		tprints(", ");
-		if (arg2 == -1)
+		if ((int) arg2 == -1)
 			tprints("PR_SET_PTRACER_ANY");
 		else
 			tprintf("%llu", arg2);
